CRUD再確認 多分確認完了

+++ app/Http/Controllers/SupportController.php
@@ -6,15 +6,16 @@ use App\Consts\FundConst;
 use App\Models\Gift;
 use App\Models\Plan;
 use App\Models\Support;
-use Illuminate\Http\Request;
+use Exception;
 use Illuminate\Support\Facades\Auth;
 
 class SupportController extends Controller
 {
     public function create(Plan $plan, Gift $gift)
     {
-        if ($plan->releseflag ) { //プロジェクトが募集しているならif文 true
-            if (!$plan->startFlag || $plan->endFlag) { //プロジェクトがすでに募集終了であるか、募集開始で基準にみたないものは支援不可
+        // dd($plan->releseFlag, $plan->startFlag, $plan->endFlag);
+        if ($plan->releseflag) {  //プロジェクトが募集しているならif文で確認をする 募集開始していないなら確認なし
+            if (!$plan->startFlag || $plan->endFlag || ($gift->limited_befor == 1)) { //プロジェクトがすでに募集終了であるか、募集開始で基準にみたないものは支援不可
                 return back()->withErrors('このプロジェクトは支援できません');
             } else {
                 //none
@@ -22,11 +23,17 @@ class SupportController extends Controller
         } else {
             //none
         }
+
         return view('plans.supports.create', compact('plan', 'gift'));
     }
 
     public function store(Plan $plan, Gift $gift)
     {
+        if ($plan->releseflag) {  //プロジェクトが募集しているならif文で確認をする 募集開始していないなら確認なし
+            if (!$plan->startFlag || $plan->endFlag || ($gift->limited_befor == 1)) { //プロジェクトがすでに募集終了であるか、募集開始で基準にみたないものは支援不可
+                return back()->withErrors('このプロジェクトは支援できません');
+            }
+        }
 
         $support = new Support();
 
@@ -53,8 +60,12 @@ class SupportController extends Controller
     public function destroy(Plan $plan, Support $support)
     {
         try {
-            $support->delete();
-        } catch (\Throwable $th) {
+            if (($plan->total - $support->money) > 0.1 || ($plan->total - $support->money) > $plan->goal) {
+                $support->delete();
+            } else {
+                throw new Exception("現在支援情報は削除できません");
+            }
+        } catch (\Exception $e) {
             return back()->withErrors('支援情報の削除に失敗しました');
         }
